# Improve StringMatcher generated code performance