--- a/src/Esprima.SourceGenerators/Helpers/SourceGenerationHelper.cs
+++ b/src/Esprima.SourceGenerators/Helpers/SourceGenerationHelper.cs
@@ -36,28 +36,6 @@ internal class StringMatcherAttribute : System.Attribute
         var indentStr = new string(' ', indent);
         var baseIndent = byLength.Length > 1 ? "        " : "    ";
 
-        static string Escape(string s)
-        {
-            return s.Replace("\"", "\\\"");
-        }
-
-        void StringEquality(StringBuilder builder, string toCheck)
-        {
-            toCheck = toCheck.Replace("\"", "\"");
-            if (!sourceIsSpan)
-            {
-                builder.Append("input == \"");
-                builder.Append(Escape(toCheck));
-                builder.Append("\"");
-            }
-            else
-            {
-                builder.Append("input.SequenceEqual(\"");
-                builder.Append(Escape(toCheck));
-                builder.Append("\".AsSpan())");
-            }
-        }
-
         if (checkNull)
         {
             sb.Append(indentStr).AppendLine("    if (input is null)");
@@ -68,6 +46,15 @@ internal class StringMatcherAttribute : System.Attribute
 
         if (byLength.Length > 1)
         {
+            var min = byLength.Min(x => x.Key);
+            var max = byLength.Max(x => x.Key);
+
+            sb.Append(indentStr).Append("    if ((uint) input.Length - ").Append(min).Append(" > ").Append(max - min).AppendLine(")");
+            sb.Append(indentStr).AppendLine("    {");
+            sb.Append(indentStr).Append("        return ").Append(returnString ? "null" : "false").AppendLine(";");
+            sb.Append(indentStr).AppendLine("    }");
+            sb.AppendLine();
+
             sb.Append(indentStr).AppendLine("    switch (input.Length)");
             sb.Append(indentStr).AppendLine("    {");
         }
@@ -83,7 +70,7 @@ internal class StringMatcherAttribute : System.Attribute
             {
                 var item = group.First();
                 sb.Append(indentStr).Append(baseIndent).Append("    return ");
-                StringEquality(sb, item);
+                StringEquality(sb, item, sourceIsSpan, startIndex: 0);
                 if (returnString)
                 {
                     sb.Append(" ? ").Append("\"").Append(Escape(item)).Append("\"").Append(" : null");
@@ -92,74 +79,46 @@ internal class StringMatcherAttribute : System.Attribute
                 continue;
             }
 
-            var discriminatorIndex = FindDiscriminatorIndex(group);
+            var discriminatorIndex = FindDiscriminatorIndex(group, 0);
 
             if (discriminatorIndex == -1)
             {
-                // hash-based or equality-based then, let compiler generate decision tree
-                var switchIndent = byLength.Length > 1 ? indentStr + indentStr + baseIndent : indentStr + baseIndent;
-                sb.Append(switchIndent).AppendLine("switch (input)");
-                sb.Append(switchIndent).AppendLine("{");
-                if (returnString)
+                // try next best effort
+                var startChars = group.Select(x => x[0]).Distinct();
+                var indentToUse = indentStr + indentStr + baseIndent;
+                if (byLength.Length > 1)
                 {
-                    foreach (var item in group)
-                    {
-                        sb.Append(switchIndent).Append("    case \"").Append(Escape(item)).AppendLine("\":");
-                        sb.Append(switchIndent).Append("        return \"").Append(Escape(item)).AppendLine("\";");
-                    }
+                    indentToUse += indentStr;
+                }
+
+                foreach (var c in startChars)
+                {
+                    sb.Append(indentToUse).Append("if (input[0] == '").Append(c).AppendLine("')");
+                    sb.Append(indentToUse).AppendLine("{");
+                    GenerateIfForStringContent(sb, group.Where(x => x[0] == c), indentToUse + indentStr, returnString, sourceIsSpan, startIndex: 1);
+                    sb.Append(indentToUse).AppendLine("}");
+                }
+
+                sb.Append(indentToUse).Append("   return ").Append(returnString ? "null" : "false").AppendLine(";");
+
+                /* alternatives
+                if (group.Key > 4)
+                {
+                    var indentToUse = byLength.Length > 1 ? indentStr + indentStr + baseIndent : indentStr + baseIndent;
+                    // hash-based or equality-based then, let compiler generate decision tree
+                    GenerateSwitchForStringContent(sb, group, indentToUse, returnString);
                 }
                 else
                 {
-                    foreach (var item in group)
-                    {
-                        sb.Append(switchIndent).Append("    case \"").Append(Escape(item)).AppendLine("\":");
-                    }
-
-                    sb.Append(switchIndent).AppendLine("        return true;");
+                    // smaller strings are fast to just brute-force check
+                    var indentToUse = byLength.Length > 1 ? indentStr + indentStr + baseIndent : indentStr + baseIndent;
+                    GenerateIfForStringContent(sb, group, indentToUse, returnString, sourceIsSpan);
                 }
-
-                sb.Append(switchIndent).AppendLine("    default:");
-                sb.Append(switchIndent).Append("        return ").Append(returnString ? "null" : "false").AppendLine(";");
-                sb.Append(switchIndent).AppendLine("}");
+                */
                 continue;
             }
 
-            sb.Append(indentStr).Append(baseIndent).Append("    switch (").Append("input[").Append(discriminatorIndex).AppendLine("])");
-            sb.Append(indentStr).Append(baseIndent).AppendLine("    {");
-
-            foreach (var item in group)
-            {
-                sb.Append(indentStr).Append(baseIndent).Append("        case '");
-                var value = item[discriminatorIndex];
-                if (value == '\'')
-                {
-                    sb.Append("\\'");
-                }
-                else
-                {
-                    sb.Append(value);
-                }
-                sb.AppendLine("':");
-
-                sb.Append(indentStr).Append(baseIndent).Append("            return ");
-                if (group.Key == 1)
-                {
-                    sb.Append("true");
-                }
-                else
-                {
-                    StringEquality(sb, item);
-                    if (returnString)
-                    {
-                        sb.Append(" ? ").Append("\"").Append(Escape(item)).Append("\"").Append(" : null");
-                    }
-                }
-                sb.AppendLine(";");
-            }
-            sb.Append(indentStr).Append(baseIndent).AppendLine("        default:");
-            sb.Append(indentStr).Append(baseIndent).Append("           return ").Append(returnString ? "null" : "false").AppendLine(";");
-
-            sb.Append(indentStr).Append(baseIndent).AppendLine("    }");
+            BuildDiscriminatorMatching(sb, group, discriminatorIndex, indent: baseIndent + indentStr, returnString, sourceIsSpan);
         }
 
         if (byLength.Length > 1)
@@ -173,10 +132,169 @@ internal class StringMatcherAttribute : System.Attribute
         return sb.ToString();
     }
 
-    private static int FindDiscriminatorIndex(IGrouping<int, string> grouping)
+    private static void BuildDiscriminatorMatching(
+        StringBuilder sb,
+        IGrouping<int, string> group,
+        int discriminatorIndex,
+        string indent,
+        bool returnString,
+        bool sourceIsSpan)
+    {
+        sb.Append(indent).Append("    switch (").Append("input[").Append(discriminatorIndex).AppendLine("])");
+        sb.Append(indent).AppendLine("    {");
+
+        foreach (var item in group)
+        {
+            sb.Append(indent).Append("        case '");
+            var value = item[discriminatorIndex];
+            if (value == '\'')
+            {
+                sb.Append("\\'");
+            }
+            else
+            {
+                sb.Append(value);
+            }
+
+            sb.AppendLine("':");
+
+            sb.Append(indent).Append("            return ");
+            if (group.Key == 1)
+            {
+                sb.Append("true");
+            }
+            else
+            {
+                StringEquality(sb, item, sourceIsSpan, startIndex: 0);
+                if (returnString)
+                {
+                    sb.Append(" ? ").Append("\"").Append(Escape(item)).Append("\"").Append(" : null");
+                }
+            }
+
+            sb.AppendLine(";");
+        }
+
+        sb.Append(indent).AppendLine("        default:");
+        sb.Append(indent).Append("           return ").Append(returnString ? "null" : "false").AppendLine(";");
+
+        sb.Append(indent).AppendLine("    }");
+    }
+
+    private static void StringEquality(StringBuilder builder, string toCheck, bool sourceIsSpan, int startIndex)
+    {
+        toCheck = toCheck.Replace("\"", "\"");
+        if (!sourceIsSpan)
+        {
+            if (startIndex != 0)
+            {
+                throw new NotImplementedException("Non-zero start index for string not implemented");
+            }
+
+            builder.Append("input == \"");
+            builder.Append(Escape(toCheck));
+            builder.Append("\"");
+        }
+        else
+        {
+            var lengthToCheck = toCheck.Length - startIndex;
+            if (lengthToCheck < 4)
+            {
+                // check char by char
+                for (var i = startIndex; i < toCheck.Length; ++i)
+                {
+                    builder.Append("input[").Append(i).Append("] == '").Append(toCheck[i]).Append("'");
+                    if (i != toCheck.Length - 1)
+                    {
+                        builder.Append(" && ");
+                    }
+                }
+            }
+            else
+            {
+                if (startIndex == 0)
+                {
+                    builder.Append("input");
+                }
+                else
+                {
+                    builder.Append("input.AsSpan(").Append(startIndex).Append(")");
+                }
+                builder.Append(".SequenceEqual(\"");
+                builder.Append(Escape(toCheck.AsSpan(startIndex).ToString()));
+                builder.Append("\".AsSpan())");
+            }
+        }
+    }
+
+    private static void GenerateSwitchForStringContent(
+        StringBuilder sb,
+        IGrouping<int, string> group,
+        string indent,
+        bool returnString)
+    {
+        sb.Append(indent).AppendLine("switch (input)");
+        sb.Append(indent).AppendLine("{");
+        if (returnString)
+        {
+            foreach (var item in group)
+            {
+                sb.Append(indent).Append("    case \"").Append(Escape(item)).AppendLine("\":");
+                sb.Append(indent).Append("        return \"").Append(Escape(item)).AppendLine("\";");
+            }
+        }
+        else
+        {
+            foreach (var item in group)
+            {
+                sb.Append(indent).Append("    case \"").Append(Escape(item)).AppendLine("\":");
+            }
+
+            sb.Append(indent).AppendLine("        return true;");
+        }
+
+        sb.Append(indent).AppendLine("    default:");
+        sb.Append(indent).Append("        return ").Append(returnString ? "null" : "false").AppendLine(";");
+        sb.Append(indent).AppendLine("}");
+    }
+
+    private static void GenerateIfForStringContent(
+        StringBuilder sb,
+        IEnumerable<string> group,
+        string indent,
+        bool returnString,
+        bool sourceIsSpan,
+        int startIndex)
+    {
+        foreach (var item in group)
+        {
+            sb.Append(indent).Append("if (");
+            StringEquality(sb, item, sourceIsSpan, startIndex);
+            sb.AppendLine(")");
+            sb.Append(indent).AppendLine("{");
+            if (returnString)
+            {
+                sb.Append(indent).Append("    return \"").Append(Escape(item)).AppendLine("\";");
+            }
+            else
+            {
+                sb.Append(indent).AppendLine("    return false;");
+            }
+            sb.Append(indent).AppendLine("}");
+        }
+
+        sb.Append(indent).Append("return ").Append(returnString ? "null" : "false").AppendLine(";");
+    }
+
+    private static string Escape(string s)
+    {
+        return s.Replace("\"", "\\\"");
+    }
+
+    private static int FindDiscriminatorIndex(IGrouping<int, string> grouping, int start)
     {
         var chars = new HashSet<char>();
-        for (var i = 0; i < grouping.Key; ++i)
+        for (var i = start; i < grouping.Key; ++i)
         {
             chars.Clear();
             var allDifferent = true;