… src/Esprima/Ast/Node.cs
… src/Esprima/Comment.cs
… src/Esprima/ParserExtensions.cs
… src/Esprima/Properties/AssemblyInfo.cs
… test/Esprima.Tests.SourceGenerators/Snapshots/StringMatcherGeneratorTests.ScannerStringMatchingGeneration.01StringMatchers.g.verified.cs