Working and testing some ideas on code-gen and improving the performance. Putting as PR so we don't do overlapping work in the area @adams85.