@lahma 2022-08-18
## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|
| Old |ParseProgram|angular-1.2.5|10.392 ms|0.0258 ms|281.2500|140.6250|4,718 KB|
| **New** |	|	| **9.922 ms (-5%)** | **0.0260 ms** | **281.2500 (0%)** | **140.6250 (0%)** | **4,718 KB (0%)** |
| Old |ParseProgram|backbone-1.1.0|1.468 ms|0.0129 ms|42.9688|17.5781|721 KB|
| **New** |	|	| **1.398 ms (-5%)** | **0.0083 ms** | **42.9688 (0%)** | **17.5781 (0%)** | **721 KB (0%)** |
| Old |ParseProgram|jquery-1.9.1|7.843 ms|0.0268 ms|234.3750|46.8750|4,079 KB|
| **New** |	|	| **7.572 ms (-3%)** | **0.0281 ms** | **242.1875 (+3%)** | **85.9375 (+83%)** | **4,079 KB (0%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|12.558 ms|0.0281 ms|375.0000|187.5000|6,304 KB|
| **New** |	|	| **12.184 ms (-3%)** | **0.0308 ms** | **375.0000 (0%)** | **187.5000 (0%)** | **6,304 KB (0%)** |
| Old |ParseProgram|mootools-1.4.5|6.477 ms|0.0213 ms|195.3125|93.7500|3,290 KB|
| **New** |	|	| **6.317 ms (-2%)** | **0.0086 ms** | **195.3125 (0%)** | **93.7500 (0%)** | **3,290 KB (0%)** |
| Old |ParseProgram|underscore-1.5.2|1.199 ms|0.0025 ms|35.1563|11.7188|603 KB|
| **New** |	|	| **1.157 ms (-4%)** | **0.0029 ms** | **35.1563 (0%)** | **11.7188 (0%)** | **603 KB (0%)** |
| Old |ParseProgram|yui-3.12.0|5.918 ms|0.0108 ms|187.5000|70.3125|3,079 KB|
| **New** |	|	| **5.766 ms (-3%)** | **0.0227 ms** | **187.5000 (0%)** | **70.3125 (0%)** | **3,079 KB (0%)** |


## Esprima.Benchmark.StringMatcherBenchmark

| **Diff**|Method|Mean|Error|Allocated|
|------- |-------|-------:|-------|-------:|
| Old |CountKeywords|8.677 ms|0.0175 ms|16 B|
| **New** |	| **7.720 ms (-11%)** | **0.0240 ms** | **10 B (-38%)** |
| Old |CountInterned|9.695 ms|0.0119 ms|19 B|
| **New** |	| **7.723 ms (-20%)** | **0.0368 ms** | **8 B (-58%)** |