@lahma 2022-08-18
I think the main work here is done, perf is better and matcher diff looks quite OK.