@lahma 2022-08-19
Tweaked styles and threshold to use indexing, also added temporary vars for chars to remove multiple indexing operations against same index entry. Got a bit faster again.

|        Method |     Mean |     Error |    StdDev | Allocated |
|-------------- |---------:|----------:|----------:|----------:|
| CountKeywords | 5.591 ms | 0.0130 ms | 0.0122 ms |       8 B |
| CountInterned | 5.694 ms | 0.0028 ms | 0.0026 ms |      10 B |